track lexer position in input to improve error message
I wanted to add this to troubleshoot another
issue I stumbled on when trying to run the
test cases (running make in the testdata dir).

Before the change.

	$ make
	go run ../cmd/minilustre <pendulum.mls >pendulum.ll
	panic: minilustre: unexpected character '.'

	goroutine 1 [running]:
	main.main()
		/home/u/goget/src/github.com/emersion/minilustre/cmd/minilustre/main.go:22 +0x234

After the change. Notice the added "at offset 60".

	$ make
	go run ../cmd/minilustre <pendulum.mls >pendulum.ll
	panic: minilustre: unexpected character '.' at offset 60

	goroutine 1 [running]:
	main.main()
		/home/u/goget/src/github.com/emersion/minilustre/cmd/minilustre/main.go:22 +0x234

From this we can tell that the lexer gets stuck
when trying to lex "0.0", which is at offset 60
of pendulum.mls.

  x = 0.0 fby (t *. dx +. x);
       ^
       |